+++ src/main/java/cz/muni/fi/pv168/project/ui/filters/ICarRideTableFilter.java
@@ -4,12 +4,11 @@ import cz.muni.fi.pv168.project.business.model.CarRide;
 import cz.muni.fi.pv168.project.business.model.Model;
 import cz.muni.fi.pv168.project.ui.filters.matchers.EntityMatcher;
 import cz.muni.fi.pv168.project.ui.model.CarRide.CarRideTableModel;
-
-import javax.swing.table.TableRowSorter;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import javax.swing.table.TableRowSorter;
 
 
 public interface ICarRideTableFilter {
@@ -17,7 +16,7 @@ public interface ICarRideTableFilter {
 
     void filterByDate(Date fromDate, Date toDate);
 
-    void filterByDistance(int fromDistance, int toDistance);
+    void filterByDistance(double fromDistance, double toDistance);
 
     void filterByPassengers(int passengersCount);
